requirements.txt
Flask==3.0.0
requests
python-dotenv
boto3==1.34.7
botocore==1.34.7
urllib3==1.26.18
zappa==0.58.0
Flask-Cors==4.0.0
svgpathtools==1.6.1
numpy==1.26.4
svgelements==1.9.6
pybind11==2.12.0
CairoSVG==2.7.1
pillow==10.3.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
python-dotenv==1.0.1
nltk==3.9